refactor: extract media provider logic
